REPORT-913: Migrate to OpenMRS Platform 2.4.0 (#262)
* Migrate to 2.4.0

* Move api-1.9

* Merge old api modules to the main

* Set the scope of reportingcompatibility to provided

(cherry picked from commit e4022f6a54ff5f4f2c0ddd0fdaf810a2f0b8d890)
