--- api-1.10/.gitignore
@@ -1,4 +1,0 @@
-/.settings
-/.classpath
-/.project
-/target
