--- api-1.9/.gitignore
@@ -1,1 +1,0 @@
-/target/
