--- api-2.0/.gitignore
@@ -1,1 +1,0 @@
-/target/
