--- api-2.2/.gitignore
@@ -1,5 +1,0 @@
-/.settings
-/.classpath
-/.project
-/target
-/bin/
