--- api-2.2/pom.xml
@@ -1,103 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>reporting</artifactId>
-		<version>1.28.0-SNAPSHOT</version>
-	</parent> 
-
-	<artifactId>reporting-api-2.2</artifactId>
-	<packaging>jar</packaging>
-	<name>reporting Module api 2.2</name>
-	<description>2.2 api project for reporting</description>
-	
-	<properties>
-		<openMRSVersion>2.2.0</openMRSVersion>
-	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>${project.parent.groupId}</groupId>
-			<artifactId>${project.parent.artifactId}-api</artifactId>
-			<version>${project.parent.version}</version>
-			<scope>provided</scope>
-		</dependency>	
-		<dependency>
-			<groupId>${project.parent.groupId}</groupId>
-			<artifactId>${project.parent.artifactId}-api-tests</artifactId>
-			<version>${project.parent.version}</version>
-			<type>test-jar</type>
-			<scope>test</scope>
-		</dependency>  
- 
-          <dependency>
-			<groupId>org.openmrs.test</groupId>
-			<artifactId>openmrs-test</artifactId>
-			<type>pom</type>
-			<scope>test</scope>
-			<version>${openMRSVersion}</version> 
-		  </dependency>  
-		<dependency>
-			<groupId>org.openmrs.api</groupId>
-			<artifactId>openmrs-api</artifactId>
-			<version>${openMRSVersion}</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>org.openmrs.api</groupId>
-			<artifactId>openmrs-api</artifactId>
-			<type>test-jar</type>
-			<scope>test</scope>
-			<version>${openMRSVersion}</version>
-		</dependency>
-
-         <dependency>
-			<groupId>org.openmrs.module</groupId>
-			<artifactId>reportingcompatibility-api</artifactId>
-			<version>${reportingCompatibilityVersion}</version>
-			<scope>test</scope>
-		</dependency>	 
-			<dependency>
-				<groupId>cglib</groupId>
-				<artifactId>cglib-nodep</artifactId>
-				<version>2.2</version>
-				<scope>test</scope>
-			</dependency> 
-			<dependency>
-    		    <groupId>org.codehaus.jackson</groupId>
-    			<artifactId>jackson-core-asl</artifactId>
-    			<version>1.9.13</version>
-    			<scope>test</scope>
-			</dependency>  
-	</dependencies>
-
-	<build>
-		<resources>
-			<resource>
-				<directory>src/main/resources</directory>
-				<filtering>true</filtering>
-			</resource>
-		</resources>
-
-		<testResources>
-			<testResource>
-				<directory>src/test/resources</directory>
-				<includes>
-					<include>**/*.properties</include>
-					<include>**/*.xml</include>
-				</includes>
-				<filtering>true</filtering>
-			</testResource>
-			<testResource>
-				<directory>src/test/resources</directory>
-				<excludes>
-					<exclude>**/*.properties</exclude>
-					<exclude>**/*.xml</exclude>
-				</excludes>
-				<filtering>false</filtering>
-			</testResource>
-		</testResources>
-	</build>
-
-</project>
