--- api-2.4/.gitignore
@@ -1,1 +1,0 @@
-/target/
